Move comparator function to dxtbx.model.compare
